chat-api/requirements.txt
openai
python-dotenv
google-genai>=1.29.0
playwright>=1.40.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0